Fix faerie fire
Faerie Fire was lowering the targets AC, making them harder to hit
instead of easier. Change the code to instead give anyone trying to
attack someone affected by faerie fire a +2 to hit.

MultiTargetedSpell was dropping the first target for charm person and
faerie fire. This may have been some special case for the first target,
but this is no longer the case.

Make anyone affected by faerie fire targetable and not -4 to hit.