Fix faerie fire
Faerie Fire was lowering the targets AC, making them harder to hit
instead of easier. Change the code to instead give anyone trying to
attack someone affected by faerie fire a +2 to hit.

MultiTargetedSpell was not drawing the missile attack animation for the
first target. Based on testing, the last target selected needs to not
have the missile attack animation played (as it is already done for the
last selected target), and the ordering of the targets should be
reversed (from last selected to first).

Make anyone affected by faerie fire targetable and not -4 to hit.